requirements.txt
 #(Keep it light.)
PyYAML>=6.0
pytest
websockets